Add usage instructions for run script

+++ run.sh
@@ -1,4 +1,10 @@
 #!/bin/sh
+if [ $# -lt 1 ] || [ "$1" = '-h' ] || [ "$1" = '--help' ]; then
+	echo "Run usage: "$0" <URL> <isLocal>"
+	echo "Example: "$0" parrot.live"
+	exit 0
+fi
+
 docker network create text-browser-network
-docker build -t text-browser . 
+docker build -t text-browser .
 docker run --rm -it --network text-browser-network text-browser npm start $1 $2
